[WPE] Platform: Fix crash on Wayland in DMA-BUF feedback handling https://bugs.webkit.org/show_bug.cgi?id=273178
Reviewed by NOBODY (OOPS!).

Move creation of the pending DMABufFeedback into the "main_device"
handler, which according to the protocol should be the first one
in the sequence. Also, move reusing the previous FormatTable into the
"done" event callback, because at that point it is certain that there
will not be more related events coming.

The rest of the changes are needed to be able to create empty
FormatTable objects for the initial creation of new DMABufFeedback
instance, and being able move-assign it later.

* Source/WebKit/WPEPlatform/wpe/wayland/WPEViewWayland.cpp:
(DMABufFeedback::FormatTable::FormatTable):
(DMABufFeedback::FormatTable::operator bool const):
(DMABufFeedback::FormatTable::operator=):
(DMABufFeedback::create):