[CMake] Allow using a system-provided libsysprof-capture package https://bugs.webkit.org/show_bug.cgi?id=275321
Reviewed by NOBODY (OOPS!).

Add a new USE_SYSTEM_SYSPROF_CAPTURE build option and a corresponding
CMake find-module. The module defines a SysProfCapture::SysProfCapture
target, the same name as already used for the copy of the library that
was already bundled under Source/ThirdParty/, therefore chosing between
the both does not need any other changes than using the find-module when
the CMake option is set.

* Source/CMakeLists.txt: Switch between the find-module for
sysprof-capture and the bundled copy depending on the value
of USE_SYSTEM_SYSPROF_CAPTURE.
* Source/cmake/FindSysProfCapture.cmake: Added.
* Source/cmake/WebKitFeatures.cmake: Add USE_SYSTEM_SYSPROF_CAPTURE
build option.