[CMake][WPE][GTK] Build a bundled copy of muprof https://bugs.webkit.org/show_bug.cgi?id=286170
Reviewed by NOBODY (OOPS!).

Import muprof under Source/ThirdParty/, and arrange CMake to build them
for the GTK and WPE ports. The ENABLE_SYSPROF_MUPROF build option may be
used to toggle compiling it.

Also add libdex, which is a dependency of muprof, with an option to
choose whether to use the system-provided version, or to statically link
the bundled copy (USE_SYSTEM_LIBDEX).

Note that while muprof is GPL-licensed, it does not get linked into
any WebKit library: it is a standalone tool. As for libdex, it is
LGPL-licensed and only linked with muprof, and both licenses are
compatible.

* Source/CMakeLists.txt: Include the muprof and lidex subdirectories in
  the build if needed.
* Source/ThirdParty/libdex/CMakeLists.txt: Added.
* Source/ThirdParty/libdex/README.WebKit: Added.
* Source/ThirdParty/libdex/cmake-config.h.in: Added.
* Source/ThirdParty/muprof/CMakeLists.txt: Added.
* Source/ThirdParty/muprof/README.WebKit: Added.
* Source/cmake/FindDex.cmake: Added.
* Source/cmake/OptionsGTK.cmake: Add ENABLE_SYSPROF_MUPROF and
  USE_SYSTEM_LIBDEX build options.
* Source/cmake/OptionsWPE.cmake: Ditto.
* Tools/Scripts/webkitpy/style/checker.py: Add style checker exceptions
  for the imported sources.