[WPEPlatform] Introduce support for Android's AHardwareBuffer https://bugs.webkit.org/show_bug.cgi?id=294317
Reviewed by NOBODY (OOPS!).

Introduce a new WPEBufferAndroid subclass of WPEBuffer that wraps an
AHardwareBuffer instance. Similarly to WPEBufferDMABuf, it supports
setting rendering and release fences (as file descriptors). The
WPEBuffer.import_to_pixels virtual method is left unimplemented for
now, but it should be possible to provide the operation for buffers
created with AHARDWAREBUFFER_USAGE_CPU_READ_{RARELY,OFTEN} in the
future.