[WPE] Log messages to the Android log where appropriate https://bugs.webkit.org/show_bug.cgi?id=295155
Reviewed by NOBODY (OOPS!).

Add an Android-specific version of the logging macros. Those use
__android_log_print() to get the messages sent to the system log
facility. LogChannel is also augmented with calls to the Android
logging functions where appropriate.

Also, on actual devices Android closes the standard error stream of
processes by default, therefore it is handy to have messages send to
the standard error stream forwarded to the system log.

In all cases, the name of the LogChannel subsystem is set as the
Android logging tag, which means that the corresponding system
property (i.e. "log.tag.<name>") needs to be set in order for logs
to be sent to the system log. This can be achieved for example with
the following command using ADB:

  adb shell setprop log.tag.WPEWebKit VERBOSE

Or, to save the setting:

  adb shell setprop persist.log.tag.WPEWebKit VERBOSE

The logging level may be set to VERBOSE, DEBUG, INFO, WARN, ERROR,
or FATAL. Those correspond to the log levels defined in <android/log.h>.
More information about how to configure the logger can be found at:

  https://android.googlesource.com/platform/system/core/+/66607ebc0e451/logd/README.property

Setting environment variables on Android is a bit of a chore, so in
the future it may be desirable to read the list of enabled logging
channels from a system property instead.

* Source/JavaScriptCore/jsc.cpp:
(main):
* Source/WTF/wtf/Assertions.cpp:
* Source/WTF/wtf/Assertions.h:
* Source/WTF/wtf/Logger.h:
(WTF::Logger::willLog const):
(WTF::Logger::log):
(WTF::Logger::logVerbose):
* Source/WTF/wtf/PlatformWPE.cmake:
* Source/WTF/wtf/WTFConfig.h:
* Source/cmake/FindAndroid.cmake: Added.
* Source/cmake/OptionsWPE.cmake: