[Cocoa] Remove unneeded WTFLogChannel::subsystem https://bugs.webkit.org/show_bug.cgi?id=295263
Reviewed by NOBODY (OOPS!).

The Cocoa ports only use the WTFLogChannel::subsystem struct member
when calling os_log_create(), and as with the other ports, the value
is the picked from the LOG_CHANNEL_WEBKIT_SUBSYSTEM macro for every
WTFLogChannel instance. Therefore, it is possible to use the macro
directly in the os_log_create() call.

As the subsystem is always picked from LOG_CHANNEL_WEBKIT_SUBSYSTEM,
this also simplifies the DEFINE_LOG_CHANNEL_[WITH_DETAILS] macros to
remove the parameter and removes the need for DEFINE_*_LOG_CHANNEL
in each component, where it is possible to use DEFINE_LOG_CHANNEL
instead.

This is a follow-up to 296858@main.

* Source/WTF/wtf/Assertions.cpp:
* Source/WTF/wtf/Assertions.h:
* Source/WTF/wtf/Logging.cpp:
* Source/WebCore/PAL/pal/Logging.cpp:
* Source/WebCore/platform/Logging.cpp:
* Source/WebDriver/Logging.cpp:
* Source/WebGPU/WebGPU/Internal/Logging.cpp:
* Source/WebKit/Platform/Logging.cpp:
* Source/WebKitLegacy/mac/Misc/WebKitLogging.m:
* Tools/TestWebKitAPI/Tests/WebCore/Logging.cpp: