[GTK][WPE] Fix non-unified builds, early 2026 edition https://bugs.webkit.org/show_bug.cgi?id=304707
Reviewed by NOBODY (OOPS!).

Most of the changes involve adding missing header inclusions. Where
possible, forward declarations are added in headers in order to prevent
adding to build time. In some instances the respective header inclusion
gets changes to the corresponding one with "...Inlines.h" suffix, as
those include the base header transitively. A few other cases involved
adding missing namespace prefixes to type names, as unified builds often
mask the issues due to "using" statements that seep in from other
translation units.

The following are the cases that were treated differently:

* Source/JavaScriptCore/runtime/TemporalPlainYearMonth.cpp:
(JSC::TemporalPlainYearMonth::addDurationToYearMonth): Deleted.
* Source/JavaScriptCore/runtime/TemporalPlainYearMonth.h:
(JSC::TemporalPlainYearMonth::addDurationToYearMonth): Moved to the
header, which was needed because the template gets used by other TUs and
hence the definition needs to be available to them.

* Source/WebCore/Modules/webxr/WebXRRenderState.cpp: Two functions where
moved to the implementation of the TU in order to avoid a circular
dependency on the WebXRWebGLLayer definition, and allow using a forward
declaration in the header.
(WebCore::WebXRRenderState::baseLayer const):
(WebCore::WebXRRenderState::setBaseLayer):
* Source/WebCore/Modules/webxr/WebXRRenderState.h:
(WebCore::WebXRRenderState::baseLayer const): Deleted.
(WebCore::WebXRRenderState::setBaseLayer): Deleted.

* Source/WebCore/bindings/js/JSSVGElementCustom.h: This needed adding
because the TU that builds the SVGElement binding needs inlines that
do not get defined by any header included by the binding generator. This
is not great as it causes the style checker to complain, but it is
possibly the less intrusive way of fixing the build issue.
* Source/WebCore/svg/SVGElement.cpp:
* Source/WebCore/svg/SVGElement.idl:

* Source/WebKit/NetworkProcess/NetworkDataTaskBlob.cpp: Removed
unused/unneeded definitions for constants.

* Source/WebKit/Platform/IPC/TransferString.cpp:
(IPC::TransferString::release): Removed template parameters.

* Source/WebKit/UIProcess/XR/openxr/OpenXRHitTestManager.cpp: Added
missing guards.
(WebKit::OpenXRHitTestManager::OpenXRHitTestManager):
* Source/WebKit/UIProcess/XR/openxr/OpenXRHitTestManager.h: Ditto.
* Source/WebKit/UIProcess/XR/openxr/OpenXRInputSource.cpp: Ditto.
* Source/WebKit/UIProcess/XR/openxr/OpenXRInputSource.h: Ditto.